Use unix timestamp (float) for selfcheck

+++ selfcheck.py
@@ -2,6 +2,6 @@ from datetime import datetime
 
 
 with open('timestamp', 'r') as timestamp:
-    updated = datetime.fromtimestamp(int(timestamp.readline()))
+    updated = datetime.fromtimestamp(float(timestamp.readline()))
     current = datetime.utcnow()
     print((current - updated).days < 8)
